refactor(@angular-devkit/build-angular): remove circular dependency in dev-server builder
This commit moves the `isEsbuildBased` function from `builder.ts` to `options.ts` to resolve a circular dependency.
