+++ goldens/circular-deps/packages.json
@@ -1,9 +1,5 @@
 [
   [
-    "packages/angular_devkit/build_angular/src/builders/dev-server/builder.ts",
-    "packages/angular_devkit/build_angular/src/builders/dev-server/options.ts"
-  ],
-  [
     "packages/angular/build/src/tools/esbuild/angular/component-stylesheets.ts",
     "packages/angular/build/src/tools/esbuild/bundler-context.ts",
     "packages/angular/build/src/tools/esbuild/utils.ts",
